test: rename acceptance test file and add TestSecondsInRadians for internal math testing
